Initial commit for "Snippets"
This note describes the initial commit for the "Snippets" feature (JDK-8201533); the commit adds JavaDoc capability to parse the `{@snippet}` tag and to process some of those parsings; more details below.

 # Detail

This note uses "standard tags" to denote tags from "Documentation Comment Specification for the Standard Doclet". The `{@snippet}` tag is not yet part of that specification.

 ## Syntax

JavaDoc parses syntax of the `{@snippet}` tag according to the following grammar:

```
              <snippet> ::= "{@snippet" <whitespace> <contents> "}"
             <contents> ::= <attributes> | <attributes> <whitespace> <body-block> | <body-block>
           <attributes> ::= <attribute> (<whitespace> <attribute>)*
            <attribute> ::= <name> <whitespace>? "=" <whitespace>? <quoted-value>
         <quoted-value> ::= "'" <single-quoted-value> "'" | '"' <double-quoted-value> '"'
           <body-block> ::= ":" <horizontal-whitespace>* <newline> <body>
           <whitespace> ::= (<horizontal-whitespace> | <vertical-whitespace>)+
<horizontal-whitespace> ::= "\t" | " "
  <vertical-whitespace> ::= "\r" | "\n" | "\f"
              <newline> ::= "\r" | "\n"
```

`<name>`, `<single-quoted-value>`, and `double-quoted-value` consist of characters that satisfy `Character.isUnicodeIdentifierPart(ch) || ch == '-'`, `ch != '\''`, and `ch != '"'` accordingly. `<body>` consists of any characters on the condition that any present `{` and `}` are balanced.

 ### Body Block

The syntax of `{@snippet}` that have `body-block` is similar (but not identical) to that of Text Blocks (JEP 378):

```
{@snippet :
...  First Line
...     Second Line
...}
```

Although `<body-block>` distinguishes between incidental and essential whitespace with `}` serving the same purpose as the closing `"""` in Text Blocks, `<body>` does not recognize escape sequences that Text Blocks recognize.

`<body-block>` is in the rightmost position of the `{@snippet}` tag and is immediately preceded by `<whitespace>`. This is because JavaDoc tags can contain `:` in their names (e.g. search for `ejb:bean` in the JDK test codebase).

`<body>` begins immediately after mandatory `<newline>` (which is itself preceded by `":" <horizontal-whitespace>*`) and ends immediately before `}`.

`<body>` is considered literal (WYSIWYG) text: when output to HTML, all special HTML symbols in are escaped so that the reader sees the `<body>` in a browser exactly the way the author saw it their editor.

`<body>` cannot contain unmatched `}` or `*/`.

 ### Attributes

The `{@snippet}` can have attributes in the form of name-value pairs with the values being single- or double-quoted:

```
{@snippet src="QuickSort.java" cut='// "Dual-pivot" version'}
```

`<single-quoted-value>` and `<double-quoted-value>` (`<*-quoted-value>` for short) begin immediately after a mandatory opening quote and ends immediately before the next occurrence of that quote.

Whitespace before and after `=` is optional. The choice between quotes exists for the author to mark `<*-quoted-value>` when `<*-quoted-value>` itself contains quotes. That is influenced by syntaxes of HTML and "shells".

Just like `<body>`, `<*-quoted-value>` is considered literal text, does not recognize escape sequences, and cannot contain `*/`. However, in contrast to `<body>`, `<*-quoted-value>` can contain an unmatched `}`.

 ### Processing

`{@snippet}` parsings that have `<body-block>` are processed into HTML as follows:

```
<pre>
body
</pre>
```
The variants without `<body-block>` are processed as if `<body>` were empty.

`{@snippet}` parsings that have `<attributes>` are processed as if `<attributes>` were empty.

 ### Notes on Syntax

1. Parts of syntax for multi-line content are borrowed from Text Blocks. This is because those parts fit; there's less to learn, less to be astonished by (Principle of Least Astonishment), and more uniformity in the Platform.
2. That said, escape sequences are not recognized in `<body>`, `<*-quoted-value>`, or anywhere else in `{@snippet}`. That is consistent with standard tags, which also do not recognize escape sequences.
3. Both `<*-quoted-value>` and `<body>` are opaque to the JavaDoc parser: it's up to an upper layer (e.g. a taglet) to interpret their content.
4. `{@snippet}` uses attributes because positional parameter-matching, which is used in standard tags, won't cut it for `{@snippet}`. This is because of the relative complexity of `{@snippet}`'s syntax. `{@snippet}` has a body block and is supposed to have at least one attribute whose value is free-form:
`cut="--- 8< --- cut here ---"`. Without more structure, the parser won't be able to distinguish between those reliably. Also, `{@snippet}` is expected to evolve in JDK as well as be extended by 3rd party, so it's expected there will be more attributes which might further complicate the matter.

 ## API

The commit adds API elements to the `com.sun.source.{doctree, util}` packages. Those API elements deliberately have empty `@since` tags: on the one hand, that doesn't suggest a certain release; on the other hand, that constantly reminds to add one. The "reminder" part works because empty `@since` tags make DocLint produce `warning: no description for @since` and are (typically) highlighted in IDEs.

I modelled snippet attributes after HTML attributes (`com.sun.source.doctree.AttributeTree`) in the separate `TagAttributeTree` type. I did NOT reuse `AttributeTree` because its specification clearly states that it's "A tree node for an attribute in an HTML element".

The reason `TagAttributeTree` is not an internal type of `SnippetTree` is that attributes are expected to be used in standard tags too, for example: `{@systemProperty scope=<label>}` (see JDK-8215040). For the same reason, relevant changes to `DocCommentParser` are not confined to just `{@snippet}`.

1. Even though attributes are expected to have unique names in the context of the `{@snippet}` tag, `SnippetTree` provides access to its attributes using `List` not `Map`.
2. In contrast to other methods in `DocTreeFactory`, `newSnippet` methods operate on trees (or collections thereof) of kinds more specific than `DocTree` (this is similar to how `newLiteralTree` and `newCodeTree` use `TextTree`).
3. That said, `TagAttributeTree` provides access to its value using `List<? extends DocTree>`, not `List<? extends TextTree>`.
4. `DocTreeFactory` provides 2 `newSnippet` methods to model optional `<body-block>`. While empty attributes are adequately denoted by an empty list, I didn't want to allow `null` (an empty `TextTree` won't cut it as absent `<body>` involves absent `<body-block>`, and thus absent `:` marker and inaccurate positions).

Items 1-3 could be revisited, should we decide that the consistency is more important.

I haven't yet decided what to do with new constants in `com.sun.source.doctree.TagAttributeTree.ValueKind`. Currently, there are only two types of quotes: `"` and `'`. In the future, there might be more quotes, e.g. `"""`. The API should provide some guidance on how to process unknown types of quotes.

Processing of `{@snippet}` is routed through `jdk.javadoc.internal.doclets.toolkit.taglets.SnippetTaglet`, and thus can be overridden by 3rd party.

 ## Testing

The commit updates some of the existing tests and adds a new one: `test/langtools/jdk/javadoc/doclet/testSnippetTag/TestSnippetTag.java`. That test verifies that `javadoc` processes syntactically correct snippets that have `<body-block>`. More extensive testing (including negative tests) will be added after the syntax will have settled.

This commit was tested as follows: `mach5 -- remote-build-and-test -j tier1,tier2,tier3`.