Fix and test "region" matching
This commit fixes two bugs and provides some tests for the "region" attribute.

The first bug was that the subsequent "stop" marker of an unrelated region was mistaken for this region's "stop" marker if this region's name was a prefix of that other region's name.

The second bug was that the region was matched eagerly: if there were multiple "stop" markers for the same region, the furthermost "stop" marker was chosen.