Read content referred from the "file" attribute
If a doc comment in a package `p` contains a snippet of the form `{@snippet file="f"}`, then the content of that snippet is sought at `p/snippet-files/f`. `snippet-files` is a directory containing package-local snippets. `snippet-files` cannot be seen as a package itself as that name has a hyphen.

Processing of the `class` attribute is disabled for now by throwing an UnsupportedOperationException.

This commit was tested as follows: `mach5 -- remote-build-and-test -j tier1,tier2,tier3`.