Add and process the "region" attribute
The "region" attribute specifies the name of a region in the content of a snippet. A region with a name "myName" consists of characters between two end-of-line comments in a snippet content:
```
// snippet-region-start:myName
...
...// snippet-region-stop:myName
```
Those comments are called markers. A region starts with the next line after opening marker and ends immediately before the closing marker. Similarly to `"""` in Text Blocks, those markers control incidental whitespace.

Content can have multiple regions. Regions can nest or overlap. If a snippet specifies the "region" attribute, then all marker comments for all affected regions are stripped before the content is output. Any empty lines created by stripping the markers are kept. If they were removed, that could interfere with line numbering which a snippet might depend on.