Add List combining methods to ClassEntry
jdk.classfile.ClassEntry and java.lang.constant.ClassDesc are two ways
of describing similar data. Often when working with Attributes, we need to
create lists that combine ClassEntrys and ClassDescs into a single list

This PR adds support working with such lists by adding 4 static methods
to ClassEntry:
* List<CE> adding(List<CE> base, List<CE> additions)
* List<CE> adding(List<CE> base, CE... additions)
* List<CE> addingSymbols(List<CE> base, List<CD> additions)
* List<CE> addingSymbols(List<CE> base, CD...additions)

This methods convert from CD to CE to create a combined List<CE>.
The methods are null-hostile in the "additions" but do not check for
nulls in the "base" List.

The returned List is mutable so these methods can be common building
blocks for composing with other CE entries.

A test has been added that validates the existing behaviours.