Use console if present, prompt and read if not
As of today, we cannot reliably say if the console represents a terminal
or not. Even if it is not, it can still be used for interaction and thus
prompt would be expected; for example in some REPLs or IDE "Run" window.

The decision is to delegate to console if it is present, and if it is
not, prompt unconditionally on stdout then read from stdin. This means
that in some cases there will be unexpected prompts; for example, if
stdin is redirected.